moved around config
